moved tests to upstream examples project
