+++ examples/jvm/src/test/scala/zio/examples/test/PredicateExampleSpec.scala
